Module tflite::model

Modules

ops

stl

Structs

AbsOptions

AbsOptionsT

AddNOptions

AddNOptionsT

AddOptions

AddOptionsT

ArgMaxOptions

ArgMaxOptionsT

ArgMinOptions

ArgMinOptionsT

BatchToSpaceNDOptions

BatchToSpaceNDOptionsT

BidirectionalSequenceLSTMOptions

BidirectionalSequenceLSTMOptionsT

BidirectionalSequenceRNNOptions

BidirectionalSequenceRNNOptionsT

Buffer

BufferT

BuiltinOptionsUnion

CallOptions

CallOptionsT

CastOptions

CastOptionsT

ConcatEmbeddingsOptions

ConcatEmbeddingsOptionsT

ConcatenationOptions

ConcatenationOptionsT

Conv2DOptions

Conv2DOptionsT

CosOptions

CosOptionsT

DepthToSpaceOptions

DepthToSpaceOptionsT

DepthwiseConv2DOptions

DepthwiseConv2DOptionsT

DequantizeOptions

DequantizeOptionsT

DivOptions

DivOptionsT

EmbeddingLookupSparseOptions

EmbeddingLookupSparseOptionsT

EqualOptions

EqualOptionsT

ExpOptions

ExpOptionsT

ExpandDimsOptions

ExpandDimsOptionsT

FakeQuantOptions

FakeQuantOptionsT

FillOptions

FillOptionsT

FlatBufferModel
An RAII object that represents a read-only tflite model, copied from disk, or mmapped. This uses flatbuffers as the serialization format.

FloorDivOptions

FloorDivOptionsT

FloorModOptions

FloorModOptionsT

FullyConnectedOptions

FullyConnectedOptionsT

GatherNdOptions

GatherNdOptionsT

GatherOptions

GatherOptionsT

GreaterEqualOptions

GreaterEqualOptionsT

GreaterOptions

GreaterOptionsT

HardSwishOptions

HardSwishOptionsT

IfOptions

IfOptionsT

Interpreter
An interpreter for a graph of nodes that input and output from tensors. Each node of the graph processes a set of input tensors and produces a set of output Tensors. All inputs/output tensors are referenced by index.

InterpreterBuilder
Build an interpreter capable of interpreting model.

L2NormOptions

L2NormOptionsT

LSHProjectionOptions

LSHProjectionOptionsT

LSTMOptions

LSTMOptionsT

LeakyReluOptions

LeakyReluOptionsT

LessEqualOptions

LessEqualOptionsT

LessOptions

LessOptionsT

LocalResponseNormalizationOptions

LocalResponseNormalizationOptionsT

LogSoftmaxOptions

LogSoftmaxOptionsT

LogicalAndOptions

LogicalAndOptionsT

LogicalNotOptions

LogicalNotOptionsT

LogicalOrOptions

LogicalOrOptionsT

MatrixDiagOptions

MatrixDiagOptionsT

MatrixSetDiagOptions

MatrixSetDiagOptionsT

MaximumMinimumOptions

MaximumMinimumOptionsT

Metadata

MetadataT

MirrorPadOptions

MirrorPadOptionsT

Model

ModelT

MulOptions

MulOptionsT

NativeTable

NegOptions

NegOptionsT

NotEqualOptions

NotEqualOptionsT

OneHotOptions

OneHotOptionsT

OpResolver
Abstract interface that returns TfLiteRegistrations given op codes or custom op names. This is the mechanism that ops being referenced in the flatbuffer model are mapped to executable function pointers (TfLiteRegistrations).

Operator

OperatorCode

OperatorCodeT

OperatorT

PackOptions

PackOptionsT

PadOptions

PadOptionsT

PadV2Options

PadV2OptionsT

Pool2DOptions

Pool2DOptionsT

PowOptions

PowOptionsT

QuantizationDetailsUnion

QuantizationParameters

QuantizationParametersT

QuantizeOptions

QuantizeOptionsT

RNNOptions

RNNOptionsT

RangeOptions

RangeOptionsT

RankOptions

RankOptionsT

ReducerOptions

ReducerOptionsT

ReshapeOptions

ReshapeOptionsT

ResizeBilinearOptions

ResizeBilinearOptionsT

ResizeNearestNeighborOptions

ResizeNearestNeighborOptionsT

ReverseSequenceOptions

ReverseSequenceOptionsT

ReverseV2Options

ReverseV2OptionsT

SVDFOptions

SVDFOptionsT

SelectOptions

SelectOptionsT

SequenceRNNOptions

SequenceRNNOptionsT

ShapeOptions

ShapeOptionsT

SkipGramOptions

SkipGramOptionsT

SliceOptions

SliceOptionsT

SoftmaxOptions

SoftmaxOptionsT

SpaceToBatchNDOptions

SpaceToBatchNDOptionsT

SpaceToDepthOptions

SpaceToDepthOptionsT

SparseToDenseOptions

SparseToDenseOptionsT

SplitOptions

SplitOptionsT

SplitVOptions

SplitVOptionsT

SquareOptions

SquareOptionsT

SquaredDifferenceOptions

SquaredDifferenceOptionsT

SqueezeOptions

SqueezeOptionsT

StridedSliceOptions

StridedSliceOptionsT

SubGraph

SubGraphT

SubOptions

SubOptionsT

Tensor

TensorT

TileOptions

TileOptionsT

TopKV2Options

TopKV2OptionsT

TransposeConvOptions

TransposeConvOptionsT

TransposeOptions

TransposeOptionsT

UnidirectionalSequenceLSTMOptions

UnidirectionalSequenceLSTMOptionsT

UniqueOptions

UniqueOptionsT

UnpackOptions

UnpackOptionsT

WhereOptions

WhereOptionsT

WhileOptions

WhileOptionsT

ZerosLikeOptions

ZerosLikeOptionsT

Enums

ActivationFunctionType

AddOptions_FlatBuffersVTableOffset

ArgMaxOptions_FlatBuffersVTableOffset

ArgMinOptions_FlatBuffersVTableOffset

BidirectionalSequenceLSTMOptions_FlatBuffersVTableOffset

BidirectionalSequenceRNNOptions_FlatBuffersVTableOffset

Buffer_FlatBuffersVTableOffset

BuiltinOperator

BuiltinOptions

CallOptions_FlatBuffersVTableOffset

CastOptions_FlatBuffersVTableOffset

CombinerType

ConcatEmbeddingsOptions_FlatBuffersVTableOffset

ConcatenationOptions_FlatBuffersVTableOffset

Conv2DOptions_FlatBuffersVTableOffset

CustomOptionsFormat

DepthToSpaceOptions_FlatBuffersVTableOffset

DepthwiseConv2DOptions_FlatBuffersVTableOffset

DivOptions_FlatBuffersVTableOffset

EmbeddingLookupSparseOptions_FlatBuffersVTableOffset

FakeQuantOptions_FlatBuffersVTableOffset

FullyConnectedOptionsWeightsFormat

FullyConnectedOptions_FlatBuffersVTableOffset

GatherOptions_FlatBuffersVTableOffset

IfOptions_FlatBuffersVTableOffset

L2NormOptions_FlatBuffersVTableOffset

LSHProjectionOptions_FlatBuffersVTableOffset

LSHProjectionType

LSTMKernelType

LSTMOptions_FlatBuffersVTableOffset

LeakyReluOptions_FlatBuffersVTableOffset

LocalResponseNormalizationOptions_FlatBuffersVTableOffset

Metadata_FlatBuffersVTableOffset

MirrorPadMode

MirrorPadOptions_FlatBuffersVTableOffset

Model_FlatBuffersVTableOffset

MulOptions_FlatBuffersVTableOffset

OneHotOptions_FlatBuffersVTableOffset

OperatorCode_FlatBuffersVTableOffset

Operator_FlatBuffersVTableOffset

PackOptions_FlatBuffersVTableOffset

Padding

Pool2DOptions_FlatBuffersVTableOffset

QuantizationDetails

QuantizationParameters_FlatBuffersVTableOffset

RNNOptions_FlatBuffersVTableOffset

ReducerOptions_FlatBuffersVTableOffset

ReshapeOptions_FlatBuffersVTableOffset

ResizeBilinearOptions_FlatBuffersVTableOffset

ResizeNearestNeighborOptions_FlatBuffersVTableOffset

ReverseSequenceOptions_FlatBuffersVTableOffset

SVDFOptions_FlatBuffersVTableOffset

SequenceRNNOptions_FlatBuffersVTableOffset

ShapeOptions_FlatBuffersVTableOffset

SkipGramOptions_FlatBuffersVTableOffset

SoftmaxOptions_FlatBuffersVTableOffset

SpaceToDepthOptions_FlatBuffersVTableOffset

SparseToDenseOptions_FlatBuffersVTableOffset

SplitOptions_FlatBuffersVTableOffset

SplitVOptions_FlatBuffersVTableOffset

SqueezeOptions_FlatBuffersVTableOffset

StridedSliceOptions_FlatBuffersVTableOffset

SubGraph_FlatBuffersVTableOffset

SubOptions_FlatBuffersVTableOffset

TensorType

Tensor_FlatBuffersVTableOffset

TransposeConvOptions_FlatBuffersVTableOffset

UnidirectionalSequenceLSTMOptions_FlatBuffersVTableOffset

UniqueOptions_FlatBuffersVTableOffset

UnpackOptions_FlatBuffersVTableOffset

WhileOptions_FlatBuffersVTableOffset

Type Definitions

AbsOptions_NativeTableType

AddNOptions_NativeTableType

AddOptions_NativeTableType

ArgMaxOptions_NativeTableType

ArgMinOptions_NativeTableType

BatchToSpaceNDOptions_NativeTableType

BidirectionalSequenceLSTMOptions_NativeTableType

BidirectionalSequenceRNNOptions_NativeTableType

Buffer_NativeTableType

CallOptions_NativeTableType

CastOptions_NativeTableType

ConcatEmbeddingsOptions_NativeTableType

ConcatenationOptions_NativeTableType

Conv2DOptions_NativeTableType

CosOptions_NativeTableType

DepthToSpaceOptions_NativeTableType

DepthwiseConv2DOptions_NativeTableType

DequantizeOptions_NativeTableType

DivOptions_NativeTableType

EmbeddingLookupSparseOptions_NativeTableType

EqualOptions_NativeTableType

ExpOptions_NativeTableType

ExpandDimsOptions_NativeTableType

FakeQuantOptions_NativeTableType

FillOptions_NativeTableType

FloorDivOptions_NativeTableType

FloorModOptions_NativeTableType

FullyConnectedOptions_NativeTableType

GatherNdOptions_NativeTableType

GatherOptions_NativeTableType

GreaterEqualOptions_NativeTableType

GreaterOptions_NativeTableType

HardSwishOptions_NativeTableType

IfOptions_NativeTableType

L2NormOptions_NativeTableType

LSHProjectionOptions_NativeTableType

LSTMOptions_NativeTableType

LeakyReluOptions_NativeTableType

LessEqualOptions_NativeTableType

LessOptions_NativeTableType

LocalResponseNormalizationOptions_NativeTableType

LogSoftmaxOptions_NativeTableType

LogicalAndOptions_NativeTableType

LogicalNotOptions_NativeTableType

LogicalOrOptions_NativeTableType

MatrixDiagOptions_NativeTableType

MatrixSetDiagOptions_NativeTableType

MaximumMinimumOptions_NativeTableType

Metadata_NativeTableType

MirrorPadOptions_NativeTableType

Model_NativeTableType

MulOptions_NativeTableType

NegOptions_NativeTableType

NotEqualOptions_NativeTableType

OneHotOptions_NativeTableType

OperatorCode_NativeTableType

Operator_NativeTableType

PackOptions_NativeTableType

PadOptions_NativeTableType

PadV2Options_NativeTableType

Pool2DOptions_NativeTableType

PowOptions_NativeTableType

QuantizationParameters_NativeTableType

QuantizeOptions_NativeTableType

RNNOptions_NativeTableType

RangeOptions_NativeTableType

RankOptions_NativeTableType

ReducerOptions_NativeTableType

ReshapeOptions_NativeTableType

ResizeBilinearOptions_NativeTableType

ResizeNearestNeighborOptions_NativeTableType

ReverseSequenceOptions_NativeTableType

ReverseV2Options_NativeTableType

SVDFOptions_NativeTableType

SelectOptions_NativeTableType

SequenceRNNOptions_NativeTableType

ShapeOptions_NativeTableType

SkipGramOptions_NativeTableType

SliceOptions_NativeTableType

SoftmaxOptions_NativeTableType

SpaceToBatchNDOptions_NativeTableType

SpaceToDepthOptions_NativeTableType

SparseToDenseOptions_NativeTableType

SplitOptions_NativeTableType

SplitVOptions_NativeTableType

SquareOptions_NativeTableType

SquaredDifferenceOptions_NativeTableType

SqueezeOptions_NativeTableType

StridedSliceOptions_NativeTableType

SubGraph_NativeTableType

SubOptions_NativeTableType

Tensor_NativeTableType

TileOptions_NativeTableType

TopKV2Options_NativeTableType

TransposeConvOptions_NativeTableType

TransposeOptions_NativeTableType

UnidirectionalSequenceLSTMOptions_NativeTableType

UniqueOptions_NativeTableType

UnpackOptions_NativeTableType

WhereOptions_NativeTableType

WhileOptions_NativeTableType

ZerosLikeOptions_NativeTableType