Struct tflite::Interpreter

pub struct Interpreter<'a, Op> where
    Op: OpResolver,  { /* fields omitted */ }

Implementations

impl<'a, Op> Interpreter<'a, Op> where
Op: OpResolver,

pub fn allocate_tensors(&mut self) -> Result<()>

Update allocations for all tensors. This will redim dependent tensors using the input tensor dimensionality as given. This is relatively expensive. If you know that your sizes are not changing, you need not call this.

pub fn print_state(&self)

Prints a dump of what tensors and what nodes are in the interpreter.

pub fn invoke(&mut self) -> Result<()>

Invoke the interpreter (run the whole graph in dependency order).

pub fn set_num_threads(&mut self, threads: c_int)

Sets the number of threads available to the interpreter threads should be >= -1 Passing in a value of -1 will let the interpreter set the number of threads available to itself.

Note that increasing the number of threads does not always speed up inference

pub fn inputs(&self) -> &[TensorIndex]

Read only access to list of inputs.

pub fn outputs(&self) -> &[TensorIndex]

Read only access to list of outputs.

pub fn variables(&self) -> &[TensorIndex]

Read only access to list of variable tensors.

pub fn tensors_size(&self) -> size_t

Return the number of tensors in the model.

pub fn nodes_size(&self) -> size_t

Return the number of ops in the model.

pub fn add_tensors(&mut self, count: size_t) -> Result<TensorIndex>

Adds count tensors, preserving pre-existing Tensor entries. Return the index of the first new tensor.

pub fn set_inputs(&mut self, inputs: &[TensorIndex]) -> Result<()>

Provide a list of tensor indexes that are inputs to the model. Each index is bound check and this modifies the consistent_ flag of the interpreter.

pub fn set_outputs(&mut self, outputs: &[TensorIndex]) -> Result<()>

Provide a list of tensor indexes that are outputs to the model Each index is bound check and this modifies the consistent_ flag of the interpreter.

pub fn set_variables(&mut self, variables: &[TensorIndex]) -> Result<()>

Provide a list of tensor indexes that are variable tensors. Each index is bound check and this modifies the consistent_ flag of the interpreter.

pub fn set_tensor_parameters_read_write(
&mut self,
tensor_index: TensorIndex,
element_type: ElementKind,
name: &str,
dims: &[usize],
quantization: QuantizationParams,
is_variable: bool
) -> Result<()>

pub fn tensor_info(&self, tensor_index: TensorIndex) -> Option<TensorInfo>

pub fn tensor_data<T>(&self, tensor_index: TensorIndex) -> Result<&[T]> where
T: ElemKindOf,

pub fn tensor_data_mut<T>(
&mut self,
tensor_index: TensorIndex
) -> Result<&mut [T]> where
T: ElemKindOf,

pub fn tensor_buffer(&self, tensor_index: TensorIndex) -> Option<&[u8]>

pub fn tensor_buffer_mut(
&mut self,
tensor_index: TensorIndex
) -> Option<&mut [u8]>